core: Move FontRenderer to font/font_renderer.rs
This makes font.rs smaller, as it's a pretty big file.
